# schemas/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "ApiErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorEnvelope",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "INVALID_ARGUMENT",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND",
          "CONFLICT",
          "RATE_LIMITED",
          "SOURCE_UNAVAILABLE",
          "PROCESSING_FAILED",
          "INTEGRITY_FAILED",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "ResponseMeta": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "const": "v1",
            "default": "v1",
            "title": "Api Version",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta",
        "type": "object"
      },
...
      "SingleResponse_SystemInfo_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SystemInfo"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SingleResponse[SystemInfo]",
        "type": "object"
      },
...
      "SystemInfo": {
        "additionalProperties": false,
        "properties": {
          "architecture": {
            "const": "modular monolith + background workers",
            "default": "modular monolith + background workers",
            "title": "Architecture",
            "type": "string"
          },
          "name": {
            "const": "AegisNews",
            "default": "AegisNews",
            "title": "Name",
            "type": "string"
          },
          "stage": {
            "const": "foundation",
            "default": "foundation",
            "title": "Stage",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version"
        ],
        "title": "SystemInfo",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Modular monolith with durable offline ingestion and canonical document retrieval.",
    "title": "AegisNews",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/system/info": {
      "get": {
        "operationId": "system_info_api_v1_system_info_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResponse_SystemInfo_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "System Info",
        "tags": [
          "system"
        ]
      }
    },
...
  },
...
}